+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## 5.6.1 - 2018-09-09
+
+* BUGFIX: dotnet restore with configfile did not work - https://github.com/fsharp/FAKE/issues/2082
+* BUGFIX: Fake cache was broken - https://github.com/fsharp/FAKE/issues/2083
+* DOCS: various improvements in the documentation - https://github.com/fsharp/FAKE/pull/2081
+
 ## 5.6.0 - 2018-09-09
 
 * NEW: Fake.DotNet.Testing.DotCover module - https://github.com/fsharp/FAKE/pull/2058